Move function file to functor directory
