+++ pynction/functors/function.py
